Add isCrossVerified func to matrixUtil
